+++ src/shaker.py
@@ -14,72 +14,113 @@ class Shaker():
 		self.yhistory = []
 		self.count = 0
 
-		# params for ShiTomasi corner detection
-		self.feature_params = dict(maxCorners = 100,
-				qualityLevel = 0.1,
-				minDistance = 7,
-				blockSize = 7 )
-
-		# Parameters for lucas kanade optical flow
-		self.lk_params = dict(winSize = (15,15),
-				maxLevel = 2,
-				criteria = (cv2.TERM_CRITERIA_EPS | cv2.TERM_CRITERIA_COUNT, 10, 0.03))
-
 		self.prev_binary = None
 		self.smoothed = np.array([])
 		self.minima = []
 		self.min_image = None
 		self.max_image = None
-	
+
+		self.start_frame = 0
+		self.num_frame = 0
+
+		self.frame_1 = None
+		self.frame_2 = None
+		self.frame_3 = None
+
+		#self.bgModel = cv2.createBackgroundSubtractorMOG2(2147483647, 0.5)
+
+	def removeBG(self, frame):
+		# apply background subtractor and erode
+		fgmask = self.bgModel.apply(frame, learningRate = 0.0000)
+		kernel = np.ones((3, 3), np.uint8)
+		fgmask = cv2.erode(fgmask, kernel, iterations=1)
+		res = cv2.bitwise_and(frame, frame, mask=fgmask)
+		return res
+
 	def local_minmax(self, arr, binary, frame):
 		num_min = 0
 		num_max = 0
-		margin = 2
+		#cnt = cv2.countNonZero(binary)
+		start_amplitude = 10
+		amplitude = 12
+		mini = 9999
+		maxi = -9999
+		find = 'start' # find min/max
 		arr = np.array(arr)#.reshape((-1,1)) # (a,1) numpy array
-		if arr.shape[0] > 20:
-			#arr = gaussian_filter(arr, sigma=7)
+		if arr.shape[0] > 3 and self.start_frame is not None:
 			arr = np.convolve(arr, [1/16,4/16,6/16,4/16,1/16], 'same')
 			arr[0] = arr[2]
 			arr[1] = arr[2]
 			arr[-1] = arr[-3]
 			arr[-2] = arr[-3]
-			for i in range(1,len(arr)-1): # minimum: slower
-				if (arr[i] < arr[i-1] - margin and arr[i] < arr[i+1]) \
-					or (arr[i] < arr[i-1] and arr[i] < arr[i+1] - margin) :
-					num_min += 1
-					if self.min_image is None:
-						self.min_image = frame
-						print('min saved')
-					self.minima.append(arr[i])
-				elif (arr[i] > arr[i-1] + margin and arr[i] > arr[i+1]) \
-					or (arr[i] > arr[i-1] and arr[i] > arr[i+1] + margin) :
-					num_max += 1
-					if self.max_image is None:
-						self.max_image = frame
-						print('max saved')
+			start = arr[self.start_frame]
+			for i in range(max(self.start_frame,0), len(arr)-2): # minimum: slower
+				#print(find)
+				if find == 'start':
+					if (arr[i] > start + start_amplitude) and arr[i-1] > arr[i]:
+						num_max += 1
+						maxi = arr[i]
+						find = 'min'
+						#print('max test {} {} {} {} {} '.format(i, arr[i], start, self.start_frame, start_amplitude))
+						if self.max_image is None:
+							self.max_image = self.frame_3
+							print('max saved: {} , frame {}'.format(maxi, i))
+					elif (arr[i] < start - start_amplitude) and arr[i-1] < arr[i]:
+						num_min += 1
+						mini = arr[i]
+						find = 'max'
+						#print('min test {} {} {} {} {} '.format(i, arr[i], start, self.start_frame, start_amplitude))
+						if self.min_image is None:
+							self.min_image = self.frame_3
+							print('min saved: {} , frame {}'.format(mini, i))
+						self.minima.append(arr[i])
+
+				elif find == 'max':
+					if (arr[i] > mini + amplitude) and arr[i-1] > arr[i]:
+						num_max += 1
+						maxi = arr[i]
+						find = 'min'
+						if self.max_image is None:
+							self.max_image = self.frame_3
+							print('max saved: {} , frame {}'.format(maxi, i))
+					elif (arr[i] < mini - amplitude*2) and arr[i-1] < arr[i]:
+						num_min += 1
+						mini = arr[i]
+						if self.min_image is None:
+							self.min_image = self.frame_3
+							print('min saved: {} , frame {}'.format(mini, i))
+						self.minima.append(arr[i])
+				elif find == 'min':
+					if (arr[i] < maxi - amplitude) and arr[i-1] < arr[i]:
+						num_min += 1
+						mini = arr[i]
+						find = 'max'
+						if self.min_image is None:
+							self.min_image = self.frame_3
+							print('min saved: {} , frame {}'.format(mini, i))
+						self.minima.append(arr[i])
+					if (arr[i] > maxi + amplitude*2) and arr[i-1] > arr[i]:
+						num_max += 1
+						maxi = arr[i]
+						if self.max_image is None:
+							self.max_image = self.frame_3
+							print('max saved: {} , frame {}'.format(maxi, i))
 			self.smoothed = arr
 			return num_min, num_max
 		return 0, 0
 
-	def visualize(self, binary, p0, p1, st):
-		mask = np.zeros_like(binary)
-		good_new = p1[st==1]
-		good_old = p0[st==1]
-		for i,(new,old) in enumerate(zip(good_new,good_old)):
-			a, b = new.ravel()
-			c, d = old.ravel()
-			mask = cv2.line(mask, (a,b),(c,d), color[i].tolist(), 2)
-			frame = cv2.circle(binary,(a,b),5,color[i].tolist(),-1)
-		img = cv2.add(frame,mask)
-		k = cv2.waitKey(10)
-		if k == 27:
-			pass
-
-	def update(self, binary):
+	def update(self, binary, frame):
 		h = binary.shape[0]
 		w = binary.shape[1]
 		weighted_y_sum = 0
 		weighted_x_sum = 0
+		ratio = 0.05
+
+		#res = self.removeBG(frame)
+		#res = cv2.cvtColor(res, cv2.COLOR_BGR2GRAY)
+		#binary = cv2.threshold(res,127,255,cv2.THRESH_BINARY)
+		#binary = np.array(binary)
+
 		idx = np.argwhere(binary > 0)
 		cnt = len(idx)
 		for i in idx:
@@ -88,6 +129,14 @@ class Shaker():
 		if cnt is not 0:
 			self.yhistory.append(weighted_y_sum/cnt)
 			self.xhistory.append(weighted_x_sum/cnt)
+		if cv2.countNonZero(binary) > binary.shape[0]*binary.shape[1]*ratio and self.start_frame is None:
+			self.start_frame = self.num_frame
+			print("start shake detection : frame "+str(self.num_frame))
+		self.num_frame += 1
+		self.frame_3 = self.frame_2
+		self.frame_2 = self.frame_1
+		self.frame_1 = frame
+
 
 	def get_minmax_image(self):
 		if self.min_image is None:
@@ -99,10 +148,11 @@ class Shaker():
 		return self.min_image, self.max_image
 
 	def shake_detect(self, binary, frame):
-		self.update(binary)
+		self.update(binary, frame)
 		num_min, num_max = self.local_minmax(self.yhistory, binary, frame)
-		print('local : ' + str(num_min) + ', ' + str(num_max))
-		if num_min >= 1 and num_max >= 2 and self.yhistory[-1] < max(self.minima) + 30:
+		if num_min + num_max is not 0: 
+			print('local : ' + str(num_min) + ', ' + str(num_max))
+		if num_min + num_max >= 4: # and self.yhistory[-1] < max(self.minima) + 15:
 			return True
 		self.prev_binary = binary
 		return False
@@ -112,7 +162,7 @@ if __name__ == "__main__":
 	cap = cv2.VideoCapture('output_bin_1.avi')
 	while cap.isOpened():
 		ret, frame = cap.read()
-		frame = cv2.resize(frame,(640, 480))
+		frame = cv2.resize(frame,(320, 240))
 		frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
 		if ret is False:
 			break
